+++ src/esri-loader.ts
@@ -12,7 +12,6 @@
 */
 const isBrowser = typeof window !== 'undefined';
 const DEFAULT_URL = 'https://js.arcgis.com/4.6/';
-
 // this is the url that is currently being, or already has loaded
 let _currentUrl;
 
@@ -22,6 +21,13 @@ function createScript(url) {
   script.src = url;
   script.setAttribute('data-esri-loader', 'loading');
   return script;
+}
+
+function createStylesheetLink(url) {
+  const link = document.createElement('link');
+  link.rel = 'stylesheet';
+  link.href = url;
+  return link;
 }
 
 // add a one-time load handler to script
@@ -60,6 +66,7 @@ function handleScriptError(script, callback) {
 // interfaces
 export interface ILoadScriptOptions {
   url?: string;
+  // TODO: css?: string;
   dojoConfig?: { [propName: string]: any };
 }
 
@@ -72,12 +79,28 @@ export const utils = {
 export function getScript() {
   return document.querySelector('script[data-esri-loader]') as HTMLScriptElement;
 }
+// TODO: export this function?
+// check if the css url has been injected or added manually
+function getCss(url) {
+  return document.querySelector(`link[href*="${url}"]`) as HTMLLinkElement;
+}
 
 // has ArcGIS API been loaded on the page yet?
 export function isLoaded() {
   const globalRequire = window['require'];
   // .on() ensures that it's Dojo's AMD loader
   return globalRequire && globalRequire.on;
+}
+
+// lazy load the CSS needed for the ArcGIS API
+export function loadCss(url) {
+  let link = getCss(url);
+  if (!link) {
+    // create & load the css library
+    link = createStylesheetLink(url);
+    document.head.appendChild(link);
+  }
+  return link;
 }
 
 // load the ArcGIS API on the page
@@ -95,7 +118,7 @@ export function loadScript(options: ILoadScriptOptions = {}): Promise<HTMLScript
       // b/c the latter will return the full url for relative paths
       const src = script.getAttribute('src');
       if (src !== options.url) {
-        // potentailly trying to load a different version of the API
+        // potentially trying to load a different version of the API
         reject(new Error(`The ArcGIS API for JavaScript is already loaded (${src}).`));
       } else {
         if (isLoaded()) {
@@ -109,7 +132,7 @@ export function loadScript(options: ILoadScriptOptions = {}): Promise<HTMLScript
     } else {
       if (isLoaded()) {
         // the API has been loaded by some other means
-        // potentailly trying to load a different version of the API
+        // potentially trying to load a different version of the API
         reject(new Error(`The ArcGIS API for JavaScript is already loaded.`));
       } else {
         // this is the first time attempting to load the API